test: split affected Git CLI journeys
